backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psutil==5.9.6
fastapi-cors==0.0.6
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
